+++ CHANGELOG.md
@@ -55,6 +55,7 @@
     * fix: aggregation can be disabled even if processed before
     * NTFS compatible abroad attachment names. Instead of a colon we use double at-sign to delimit country code from the chosen e-mail when splitting by an incident-contact field.
     * fix: PickBase used from CLI through --external
+    * fix: empty `convey.log` is no more created in the directory when writing `python3 -m [LETTER].[TAB]` into terminal
 
 ## 1.3.1 (2020-01-31)
 * thread processing